@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HamDotNetToolkit
+namespace HamDotNetToolkit.Models
 {
     public struct LatLong
     {
-            public double Latitude;
-            public double Longitude;
+        public double Latitude;
+        public double Longitude;
     }
 }
